Separate DR resolution components
Remove unnecessary formik usage
