+++ frontend/src/components/single-contract/DisputeJobFormButton.tsx
@@ -1,6 +1,5 @@
 import { useJob } from 'components/smart-contracts/hooks/useJob';
 import { DisputeJobModal } from 'components/smart-contracts/modals/DisputeJobModal';
-import { Form, Formik } from 'formik';
 import React, { useState } from 'react';
 import { useHistory } from 'react-router-dom';
 
@@ -13,48 +12,38 @@ export const DisputeJobFormButton: React.FC<IDisputeJobFormButtonProps> = (
 ) => {
   const { label } = props;
   const { job, isLoading } = useJob();
-  const [showDisputeJobModal, setShowDisputeJobModal] = useState(false);
   const history = useHistory();
 
-  return job && !isLoading ? (
-    <Formik
-      initialValues={{}}
-      onSubmit={(values) => {
-        setShowDisputeJobModal(true);
-      }}
-    >
-      {({ isSubmitting, setSubmitting }) => (
-        <Form style={isSubmitting ? { opacity: 0.5 } : {}}>
-          <button
-            disabled={isSubmitting}
-            type="submit"
-            className={
-              'focus:outline-none inline-flex items-center rounded-md border border-transparent px-4 py-2 shadow-sm focus:ring-2 focus:ring-indigo-500 focus:ring-offset-2' +
-              (isSubmitting
-                ? ' bg-indigo-100 hover:bg-indigo-100'
-                : ' bg-indigo-600 hover:bg-indigo-700')
-            }
-          >
-            {label || 'Close Job'}
-          </button>
+  const [showDisputeJobModal, setShowDisputeJobModal] = useState(false);
 
-          {isSubmitting && (
-            <DisputeJobModal
-              jobId={job.id}
-              show={showDisputeJobModal}
-              onConfirmed={(jobId: string) => {
-                setShowDisputeJobModal(false);
-                // reload
-                history.push(`/contract/${jobId}`);
-              }}
-              onError={() => {
-                setSubmitting(false);
-                setShowDisputeJobModal(false);
-              }}
-            />
-          )}
-        </Form>
-      )}
-    </Formik>
+  return job && !isLoading ? (
+    <div style={showDisputeJobModal ? { opacity: 0.5 } : {}}>
+      <button
+        disabled={showDisputeJobModal}
+        onClick={() => setShowDisputeJobModal(true)}
+        type="submit"
+        className={
+          'focus:outline-none inline-flex items-center rounded-md border border-transparent px-4 py-2 shadow-sm focus:ring-2 focus:ring-red-500 focus:ring-offset-2' +
+          (showDisputeJobModal
+            ? ' bg-red-100 hover:bg-red-100'
+            : ' bg-red-600 hover:bg-red-700')
+        }
+      >
+        {label || 'Close Job'}
+      </button>
+
+      <DisputeJobModal
+        jobId={job.id}
+        show={showDisputeJobModal}
+        onConfirmed={(jobId: string) => {
+          setShowDisputeJobModal(false);
+          // reload
+          history.push(`/contract/${jobId}`);
+        }}
+        onError={() => {
+          setShowDisputeJobModal(false);
+        }}
+      />
+    </div>
   ) : null;
 };
